+++ src/client/index.ts
@@ -1,8 +1,12 @@
+import { error, info, initLogger, loggerAvailable } from "../logging";
+import { getEnvironmentMode } from "../environment";
+import { initI18n } from "../i18n";
+import { createCommands } from "../commands";
+
 import { Client, Collection, Events, GatewayIntentBits } from "discord.js";
 
 import { CustomCommand } from "../@types/CustomCommand";
 import { replacements } from "../replacements";
-import { error, info } from "../logging";
 
 export function createClient(commands: CustomCommand[]): Client {
   const replacementsEntries = Object.entries(replacements);
@@ -103,3 +107,28 @@ export function createClient(commands: CustomCommand[]): Client {
 
   return client;
 }
+
+async function main(): Promise<void> {
+  const environmentMode = getEnvironmentMode();
+  initLogger(environmentMode, "bot");
+
+  const locale = process.env.LOCALE ?? "";
+  await initI18n(locale);
+
+  const commands = createCommands();
+  const client = createClient(commands);
+
+  await client.login(process.env.DISCORD_BOT_TOKEN);
+}
+
+main()
+  .then()
+  .catch((e: Error) => {
+    if (loggerAvailable()) {
+      // TODO: Refactor `error()` parameters to accept an `Error` object
+      error(`Exception thrown from main:\n ${e.name}: ${e.message}!`);
+      return;
+    }
+
+    console.error(`Exception thrown from main:\n ${e.name}: ${e.message}!`);
+  });
