+++ src/index.ts
@@ -1,34 +1,19 @@
 import dotenv from "dotenv";
 
-import { error, initLogger, loggerAvailable } from "./logging";
-import { initI18n } from "./i18n";
-import { createCommands } from "./commands";
-import { createClient } from "./client";
+import { info, initLogger } from "./logging";
 import { getEnvironmentMode } from "./environment";
 
-async function main(): Promise<void> {
-  dotenv.config();
+import { ShardingManager } from "discord.js";
 
-  const environmentMode = getEnvironmentMode();
-  initLogger(environmentMode, "bot");
+dotenv.config();
 
-  const locale = process.env.LOCALE ?? "";
-  await initI18n(locale);
+const manager = new ShardingManager("./dist/client/index.js", { token: process.env.DISCORD_BOT_TOKEN });
 
-  const commands = createCommands();
-  const client = createClient(commands);
+const environmentMode = getEnvironmentMode();
+initLogger(environmentMode, "bot");
 
-  await client.login(process.env.DISCORD_BOT_TOKEN);
-}
+manager.on('shardCreate', (shard) => {
+  info(`Launched shard ${shard.id}`);
+});
 
-main()
-  .then()
-  .catch((e: Error) => {
-    if (loggerAvailable()) {
-      // TODO: Refactor `error()` parameters to accept an `Error` object
-      error(`Exception thrown from main:\n ${e.name}: ${e.message}!`);
-      return;
-    }
-
-    console.error(`Exception thrown from main:\n ${e.name}: ${e.message}!`);
-  });
+void manager.spawn();
